pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Rafael Ortiz <rafaelortiz99@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
